graph, chain/ethereum: Move entity conversation traits
Move these into `graph` so they can be used in other places as
well (like other chain integrations in the future).
